подключения нет, мапперы кривые

+++ src/main/java/by/toukachmikhail/taskmanagementsystem/controllers/impl/UserControllerImpl.java
@@ -1,33 +1,33 @@
 package by.toukachmikhail.taskmanagementsystem.controllers.impl;
 
-import by.toukachmikhail.taskmanagementsystem.controllers.UserTaskController;
-import by.toukachmikhail.taskmanagementsystem.dto.TaskDto;
+import by.toukachmikhail.taskmanagementsystem.controllers.UserController;
 import by.toukachmikhail.taskmanagementsystem.dto.UserDto;
 import by.toukachmikhail.taskmanagementsystem.services.UserService;
 import java.util.List;
+import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/api/v1/task-management/users")
+@RequestMapping("/api/v1/users")
 @RequiredArgsConstructor
-public class UserControllerImpl implements UserTaskController{
+public class UserControllerImpl implements UserController {
 
   private final UserService userService;
 
   @Override
-  @GetMapping("/")
-  public ResponseEntity<List<UserDto>> showAllUsers() {
+  @GetMapping()
+  public ResponseEntity<List<UserDto>> getAllUsers() {
     List<UserDto> usersDtoList = userService.getAllUsers();
     return ResponseEntity.ok()
         .contentType(MediaType.APPLICATION_JSON)
@@ -36,44 +36,35 @@ public class UserControllerImpl implements UserTaskController{
 
   @Override
   @GetMapping("/{user_id}")
-  public ResponseEntity<UserDto> showSingleUser(@PathVariable("user_id") Long taskId) {
+  public ResponseEntity<UserDto> getUserById(@PathVariable("user_id") Long userId) {
 
-    UserDto userDto = userService.getUser(taskId);
-
-    return ResponseEntity.ok()
-        .contentType(MediaType.APPLICATION_JSON)
-        .body(userDto);
+    Optional<UserDto> userDto = userService.getUserById(userId);
+    return userDto.map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.notFound().build());
   }
 
   @Override
-  @PostMapping()
-  public ResponseEntity<HttpStatus> addNewTask(@RequestBody UserDto userDto){
-
-    userService.saveUser(userDto);
-    return ResponseEntity.ok(HttpStatus.OK);
+  @PostMapping
+  public ResponseEntity<UserDto> createUser(@RequestBody UserDto userDto) {
+    return ResponseEntity.status(HttpStatus.CREATED).body(userService.saveUser(userDto));
   }
 
   @Override
-  @PatchMapping("/")
-  public ResponseEntity<HttpStatus> updateTask(@RequestBody TaskDto taskDto) {
-
-    userService.correctUsersInfo(taskDto);
-    return ResponseEntity.ok(HttpStatus.OK);
+  @PutMapping("/{userId}")
+  public ResponseEntity<UserDto> updateUser(@PathVariable Long userId,
+      @RequestBody UserDto userDto) {
+    Optional<UserDto> updatedUserDto = userService.updateUser(userId, userDto);
+    return updatedUserDto.map(ResponseEntity::ok)
+        .orElseGet(() -> ResponseEntity.notFound().build());
   }
 
   @Override
-  @DeleteMapping()
-  public ResponseEntity<HttpStatus> removeAllTasks(){
-
-    userService.deleteAllUsers();
-    return ResponseEntity.ok(HttpStatus.OK);
-  }
-
-  @Override
-  @DeleteMapping("/{user_id}")
-  public ResponseEntity<HttpStatus> updateTask(@PathVariable("user_id") Long userId){
-
-    userService.deleteUser(userId);
-    return ResponseEntity.ok(HttpStatus.OK);
+  @DeleteMapping("/{userId}")
+  public ResponseEntity<Void> deleteUser(@PathVariable Long userId) {
+    if (userService.getUserById(userId).isPresent()) {
+      userService.deleteUser(userId);
+      return ResponseEntity.noContent().build();
+    } else {
+      return ResponseEntity.notFound().build();
+    }
   }
 }
